Condense creation of service deployment pipelines into a single loop
The idempotent nature of the ensure_* methods is key here. This change also enables the function generating the migration rollback job to launch an EC2 instance when necessary. Previously, the function assumed that an instance was already running. This is problematic when jobs are responsible for managing their own instances (i.e., both launching them and tearing them down, when necessary).

ECOM-6438